+++ packages/api-client/src/store/store-context.ts
@@ -1,4 +1,3 @@
-import type { RouterPathParams } from '@/store/router-params'
 import type { Cookie } from '@scalar/oas-utils/entities/cookie'
 import type { Environment } from '@scalar/oas-utils/entities/environment'
 import type {
@@ -11,11 +10,8 @@ import type {
 } from '@scalar/oas-utils/entities/spec'
 import type { Workspace } from '@scalar/oas-utils/entities/workspace'
 import type { Mutators } from '@scalar/object-utils/mutator-record'
-import type { ComputedRef } from 'vue'
 
 export type StoreContext = {
-  routerParams: ComputedRef<RouterPathParams>
-  //
   collections: Record<string, Collection>
   collectionMutators: Mutators<Collection>
   //
