+++ packages/api-client/src/views/Request/RequestSection/RequestSection.vue
@@ -39,7 +39,7 @@ const sections = computed(() => {
 
 // If security = [] or [{}] just hide it on readOnly mode
 const isAuthHidden = computed(
-  () => isReadOnly.value && activeRequest.value?.security?.length === 0,
+  () => isReadOnly && activeRequest.value?.security?.length === 0,
 )
 
 type ActiveSections = (typeof sections.value)[number]
