+++ packages/oas-utils/src/entities/workspace/workspace.ts
@@ -33,8 +33,6 @@ export const workspaceSchema = z.object({
   name: z.string().default('Default Workspace'),
   /** Workspace description */
   description: z.string().default('Basic Scalar Workspace'),
-  /** Controls read only mode for most entitites, but not things like params */
-  isReadOnly: z.boolean().default(false),
   /** List of all collection uids in a given workspace */
   collections: z.array(z.string()).default([]),
   /** List of all environment uids in a given workspace */
